% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_netscanr_db}
\alias{create_netscanr_db}
\title{Create netscanR SQLite Database}
\usage{
create_netscanr_db(file)
}
\arguments{
\item{file}{path to the SQLite database. If the file does not exist, it is
created. The file should have the ending \code{.sqlite}. If omitted, the ending
is added automatically.}
}
\value{
invisible \code{TRUE} if the database and the table were created successfully.
}
\description{
Create a SQLite database that can be used to archive the results of
network scans with \code{\link[=run_arp_scan]{run_arp_scan()}}.
}
\details{
This function creates an SQLite database that can be used to store repeated
network scans. If the file does not exist, it is created. A table \code{netscanr}
is created. If this table already exists, the function aborts with an
error.

The table \code{netscanr} has all the columns of the tibble created by
\code{\link[=run_arp_scan]{run_arp_scan()}} plus an additional column \code{timestamp} that stores the
time when the scan was performed.
}
