% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifconfig.R
\name{find_ifconfig}
\alias{find_ifconfig}
\title{Find the ifconfig Executable}
\usage{
find_ifconfig()
}
\value{
The path to the ifconfig executable as a character.
If it is not found, an empty character is returned.
}
\description{
Find the path to the ifconfig executable.
}
