% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{find_ping}
\alias{find_ping}
\title{Find the ping Executable}
\usage{
find_ping()
}
\value{
The path to the ping executable as a character.
If it is not found, an empty character is returned.
}
\description{
Find the path to the ping executable.
}
