% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_mac}
\alias{is_mac}
\alias{is_ip}
\title{Check Validity of MAC and IP addresses}
\usage{
is_mac(x)

is_ip(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a logical vector of the same length as \code{x}
}
\description{
Check Validity of MAC and IP addresses
}
