% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifconfig.R
\name{list_interfaces}
\alias{list_interfaces}
\title{List Available Network Interfaces}
\usage{
list_interfaces()
}
\value{
a tibble with the following columns:
\itemize{
\item \code{name}: the name of the interface
\item \code{mtu}: the maximum transmission unit, the size in bytes of the largest
packet that can be transmitted
\item \code{mac}: the MAC address of the interface
\item \code{ipv4}: the IPv4 address that is currently assigned to the interface.
This is \code{NA} if the interface is not connected.
}
}
\description{
Use \code{ifconfig} to list available network interfaces.
}
\details{
\code{ifconfig} must be available on the system for this function to work.
Use \code{\link[=find_ifconfig]{find_ifconfig()}} to check whether this is the case.
}
