% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netscanR.R
\docType{package}
\name{netscanR-package}
\alias{netscanR}
\alias{netscanR-package}
\title{netscanR: Scan the Network With arp-scan and Parse the Results}
\description{
Scan the network with arp-scan and parse the results into a tibble.
}
\keyword{internal}
