% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Ping IP addresses}
\usage{
ping(hosts, count = 3, interval = 0.1, timeout = 1)
}
\arguments{
\item{hosts}{a character vector of IP addresses.}

\item{count}{integer giving the number of packets to send.}

\item{interval}{numeric giving the time interval between packets.}

\item{timeout}{numeric giving the time to wait for a response.}
}
\value{
a named logical vector of the same length as \code{hosts} indicating whether the
IP address could be reached.
}
\description{
Ping a list of IP addresses and return for each one, whether it could be
reached.
}
\details{
\code{ping()} sends \code{count} packets to each IP address in \code{hosts}. It returns
\code{TRUE}, if at least one response was received and \code{FALSE} otherwise.
}
