% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_list.R
\name{read_device_list}
\alias{read_device_list}
\title{Read List of Network Devices}
\usage{
read_device_list(file)
}
\arguments{
\item{file}{the path to the file}
}
\value{
a tibble with the columns \code{mac}, \code{description}, and \code{ip}.
}
\description{
Read a list of network devices from a csv file. This list can be used to
enrich the table produced by \code{\link[=run_arp_scan]{run_arp_scan()}} with additional information
and detect unknown devices in the network.
}
\details{
The file containing the device list must be a csv file separated by commas
(","). The first row must contain column headers. The supported columns
are:
\itemize{
\item \code{mac}: The MAC address of a device
\item \code{description}: A description of the host. This can be any string and does
not need to be equal to the host name. The description will be used to
uniquely identify the device when the device list is used.
\item \code{ip}: Optional. The expected IP address of the device (if any). The column
can be omitted as a whole (in which case it will be filled with \code{NA}) or
only some entries may be omitted for devices that have no expected
(i.e., fixed) IP address.
}

The package contains a valid example file. The path to this file can be
obtained by

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("extdata", "device_list.csv", package = "netscanR")
}\if{html}{\out{</div>}}
}
