% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{read_netscanr_db}
\alias{read_netscanr_db}
\title{Read netscanR SQLite Database}
\usage{
read_netscanr_db(file, start = NULL, end = NULL, tz = "")
}
\arguments{
\item{file}{path to the SQLite database.}

\item{start, end}{start and end timestamp for the data to be read.}

\item{tz}{time zone to be used for the timestamps. If omitted, the default
time zone of the system is used.}
}
\value{
a tibble with the data read from the database. It has all the columns
of the tibble created by \code{\link[=run_arp_scan]{run_arp_scan()}} plus an additional column
\code{timestamp} with the time when the scan was performed.
}
\description{
Read data from a netscanR SQLite database.
}
