% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arp_scan.R
\name{run_arp_scan}
\alias{run_arp_scan}
\title{Scan the Network with \code{arp-scan}}
\usage{
run_arp_scan(localnet = TRUE, interface = NULL, hosts = NULL, verbose = FALSE)
}
\arguments{
\item{localnet}{logical, should the local network be scanned by adding
the argument \code{--localnet} to the command? This argument is ignored if
\code{hosts} is provided.}

\item{interface}{character, the network interface to use for the scan.
If not specified, \code{arp-scan} will search the system interface list for
the lowest numbered, configured up interface (excluding loopback).}

\item{hosts}{character, the hosts to scan. They can be provided as IP
addresses or hostnames. A vector of multiple single hosts can be passed
or a range of hosts can be specified in one of three formats:
CIDR notation (e.g. "192.168.1.0/24"), a range of IP addresses indicated
by a dash (e.g. "192.168.1.1-192.168.1.10") or an IP address with mask
(e.g. "192.168.1.0:255.255.255.0").
If \code{hosts} is provided, \code{localnet} is ignored.}

\item{verbose}{logical, should additional output be printed to the console?}
}
\value{
a tibble with the following columns:
\itemize{
\item \code{interface}: the name of the interface that was scanned
\item \code{ip}: the IP address of the host
\item \code{mac}: the MAC address of the host
\item \code{vendor}: the vendor of the network card, which may differ from the
manufacturer of the device
}
}
\description{
Scan the Network with \code{arp-scan}
}
\details{
\code{arp-scan} must be installed on the system for this function to work. Use
\code{\link[=find_arp_scan]{find_arp_scan()}} to check whether this is the case. Also, root privileges
are required to run \code{arp-scan}. So, either R must be run as root or the
setuid must be set on the \code{arp-scan} executable (see the documentation of
\code{\link[=can_run_arp_scan]{can_run_arp_scan()}} for more information). Use \code{can_run_arp_scan()} to
check whether the current user has the necessary privileges.
}
