% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_list.R
\name{update_device_list}
\alias{update_device_list}
\title{Update List of Network Devices with Results from arp-scan}
\usage{
update_device_list(
  arp_scan_table,
  device_list_file,
  vendor_to_description = TRUE,
  update_ip = FALSE,
  verbose = interactive()
)
}
\arguments{
\item{arp_scan_table}{a tibble with the output from \code{\link[=run_arp_scan]{run_arp_scan()}}}

\item{device_list_file}{path to a device list file. If it exists, missing devices
are added. If it does not exist, a new device list is created containing
all the devices from \code{arp_scan_table}.}

\item{vendor_to_description}{logical. If \code{TRUE}, the vendor information is
used as the description of the device. If \code{FALSE}, the description is left
empty.}

\item{update_ip}{logical. Should the expected IP in the device list be
updated with the IP addresses from \code{arp_scan_table}.}

\item{verbose}{should additional output be produced. This defaults to
\code{TRUE} in interactive sessions.}
}
\value{
a tibble containing the updated device list. The file \code{device_list} is
updated or created as a side effect.
}
\description{
Use the results from arp-scan to update the device list.
}
