% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{update_netscanr_db}
\alias{update_netscanr_db}
\title{Update netscanR SQLite Database}
\usage{
update_netscanr_db(file, device_list_file, update_known = FALSE)
}
\arguments{
\item{file}{path to the SQLite database.}

\item{device_list_file}{path to a csv file with a list of devices
(see \code{\link[=read_device_list]{read_device_list()}}).}

\item{update_known}{logical. Should also the descriptions of known devices
be updated? By default, only devices that were not known at the time the
database was written, i.e., that were not in the device list used at that
moment, will be updated. Setting this to \code{TRUE} will update the descriptions
of all appearances of a device in the database.}
}
\value{
a data frame with the devices for which the description was updated for at
least one timestamp. As a side effect, the database is updated.
}
\description{
Update the descriptions in a netscanR SQLite database using a device list.
}
\details{
This function rewrites the description of unknown devices in a netscanR
SQLite database. This can be used to retrospectively mark devices as known
and add a description to them.

By setting \code{update_known = TRUE}, it can also update existing descriptions
of known devices to a new value.

Note that this reads and writes the entire database such that it may be slow
for large databases.
}
