% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{write_netscanr_db}
\alias{write_netscanr_db}
\title{Write to netscanR SQLite Database}
\usage{
write_netscanr_db(arp_scan_result, file, timestamp = Sys.time())
}
\arguments{
\item{arp_scan_result}{tibble with the results of \code{\link[=run_arp_scan]{run_arp_scan()}}.}

\item{file}{path to the SQLite database created with \code{\link[=create_netscanr_db]{create_netscanr_db()}}.}

\item{timestamp}{timestamp to be written to the database for this scan.
If omitted, the current time is used.}
}
\value{
the data written to the database, invisibly.
}
\description{
Write the output of a network scan with \code{\link[=run_arp_scan]{run_arp_scan()}} to an netscanR
SQLite database.
}
