% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netscanR.R
\docType{package}
\name{netscanR-package}
\alias{netscanR}
\alias{netscanR-package}
\title{netscanR: Scan the Local Network with R}
\description{
netscanR allows you to scan your local network using \code{arp-scan}.
Output from \code{arp-scan} is parsed into a tibble that can be analysed using R.
}
\keyword{internal}
