% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_arp_scan.R
\name{run_arp_scan}
\alias{run_arp_scan}
\title{Scan the Network with \code{arp-scan}}
\usage{
run_arp_scan(
  localnet = TRUE,
  interface = NULL,
  hosts = NULL,
  device_list = NULL,
  repeater = c(),
  retry = 2,
  interval = 2000,
  require_ping = 0,
  verbose = FALSE
)
}
\arguments{
\item{localnet}{logical, should the local network be scanned by adding
the argument \code{--localnet} to the command? This argument is ignored if
\code{hosts} is provided.}

\item{interface}{character, the network interface to use for the scan.
If not specified, \code{arp-scan} will search the system interface list for
the lowest numbered, configured up interface (excluding loopback).}

\item{hosts}{character, the hosts to scan. They can be provided as IP
addresses or hostnames. A vector of multiple single hosts can be passed
or a range of hosts can be specified in one of three formats:
CIDR notation (e.g. "192.168.1.0/24"), a range of IP addresses indicated
by a dash (e.g. "192.168.1.1-192.168.1.10") or an IP address with mask
(e.g. "192.168.1.0:255.255.255.0").
If \code{hosts} is provided, \code{localnet} is ignored.}

\item{device_list}{character giving the path where the device list is
stored as a csv-file or a tibble containing a device list. See
\code{\link[=read_device_list]{read_device_list()}} for more information.}

\item{repeater}{character, the mac address(es) of one or several repeaters
that may be present in the network. \code{arp-scan} reports devices that are
connected to a repeater with the mac address of the repeater such that they
cannot be correctly identified. (Note that this mac address may even deviate
from the mac address that is reported by \code{arp-scan} for the repeater itself.)
If one  or multiple repeater addresses are provided, the devices
connected to the repeater(s) are identified via their IP address instead.
Of course, this only works, if the IP addresses are static.}

\item{retry}{integer, the number of times to retry sending the ARP request.
Using more retries leads to a more reliable detection of hosts, but also
increases the time the scan takes.}

\item{interval}{integer, the time interval in microseconds between ARP
requests. Setting this to a lower value speeds up the scan but may result
in an ARP storm which can disrupt network operation. Also, setting the
interval too low can send packets faster than the network interface can
transmit them, which will eventually fill the kernel's transmit buffer
resulting in the error message: No buffer space available.}

\item{require_ping}{integer, the number of pings to send to each host to
check that it is reachable. If a host does not respond to at least one ping,
it is removed from the output. Set to 0 to not require any successful pings.}

\item{verbose}{logical, should additional output be printed to the console?}
}
\value{
a tibble with the following columns:
\itemize{
\item \code{interface}: the name of the interface that was scanned
\item \code{ip}: the IP address of the host
\item \code{mac}: the MAC address of the host
\item \code{vendor}: the vendor of the network card, which may differ from the
manufacturer of the device
}

If \code{device_list} is provided, the tibble contains three additional columns:
\itemize{
\item \code{description}: the description associated with the device from the
device list. If no description is provided or if the device is not listed
in the device list, this column is \code{NA}.
\item \code{expected_ip}: a logical indicating whether the IP address of the device
corresponds to the expected address given in the device list. If no
IP is provided in the device list, this column is \code{NA}.
\item \code{known_device}: a logical indicating whether the device is known, i.e.,
whether its MAC address is contained in the device list.
}
}
\description{
Scan the Network with \code{arp-scan}
}
\details{
\code{arp-scan} must be installed on the system for this function to work. Use
\code{\link[=find_arp_scan]{find_arp_scan()}} to check whether this is the case. Also, root privileges
are required to run \code{arp-scan}. So, either R must be run as root or the
setuid must be set on the \code{arp-scan} executable (see the documentation of
\code{\link[=can_run_arp_scan]{can_run_arp_scan()}} for more information). Use \code{can_run_arp_scan()} to
check whether the current user has the necessary privileges.
}
